gin-web-server/requirements.txt
rembg==2.0.50
pillow>=8.3.0
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.18.0
